+++ packages/core/src/util/index.js
@@ -3,6 +3,7 @@
 module.exports = exports = {
   applicationUnderMonitoring: require('./applicationUnderMonitoring'),
   atMostOnce: require('./atMostOnce'),
+  buffer: require('./buffer'),
   clone: require('./clone'),
   compression: require('./compression'),
   hasThePackageBeenInitializedTooLate: require('./initializedTooLateHeuristic'),